♻️ Move *runner.go to command package
