+++ command/runner.go
@@ -1,0 +1,5 @@
+package command
+
+type Runner interface {
+	Run(c *Command) error
+}
