+++ command/tmux_runner.go
@@ -1,4 +1,4 @@
-package main
+package command
 
 import (
 	"bytes"
@@ -11,8 +11,6 @@ import (
 	"regexp"
 	"strings"
 	"time"
-
-	"github.com/akm/git-exec/command"
 )
 
 type TmuxRunner struct {
@@ -27,7 +25,7 @@ type TmuxRunner struct {
 
 var _ Runner = (*TmuxRunner)(nil)
 
-func newTmuxRunner(debugLog bool) *TmuxRunner {
+func NewTmuxRunner(debugLog bool) *TmuxRunner {
 	doneStringPrefix := "git-exec-done"
 	return &TmuxRunner{
 		session:        "git-exec-session",
@@ -39,7 +37,7 @@ func newTmuxRunner(debugLog bool) *TmuxRunner {
 	}
 }
 
-func (x *TmuxRunner) Run(c *command.Command) (rerr error) {
+func (x *TmuxRunner) Run(c *Command) (rerr error) {
 	if _, err := exec.LookPath("tmux"); err != nil {
 		return fmt.Errorf("tmux is not installed. Please install tmux. See https://github.com/tmux/tmux/wiki/Installing")
 	}
