--- runner.go
@@ -1,9 +1,0 @@
-package main
-
-import (
-	"github.com/akm/git-exec/command"
-)
-
-type Runner interface {
-	Run(c *command.Command) error
-}
